feat(quiz): 퀴즈에 대한 문제 정답 제출 api 추가, 기존 퀴즈 완료 api 삭제, questionDto의 정답 필드 제거

+++ src/main/java/doritos/doriroom/quiz/domain/Question.java
@@ -33,7 +33,7 @@ public class Question {
     private String option4;
 
     @Column(nullable = false)
-    private int correctAnswer; // 정답 번호 (1, 2, 3, 4 중)
+    private byte correctAnswer; // 정답 번호 (1, 2, 3, 4 중)
 
     @Column(nullable = false, length = 1000)
     private String commentary; // 문제 해설 내용
